requirements.txt
streamlit
qdrant-client
torch
torchvision
pillow
requests
pandas
numpy
tqdm